requirements.txt
numpy>=1.14.5
keras==2.2.4
tensorflow==1.10.1
plaidml>=0.6.2
pillow>=4.0.0
tqdm>=4.19.4